feat(Yonas.implement-authentication) tests added

+++ Backend/starter_project_zsilentcoders/BlogApp3/BlogApp.API/Controllers/AuthController.cs
@@ -8,13 +8,13 @@ using BlogApp.Application.Features.Users.DTOs;
 using BlogApp.Application.Responses;
 using MediatR;
 using AutoMapper;
-
+using BlogApp.Api.Controllers;
 
 namespace BlogApp.API.Controllers;
 
 [ApiController]
 [Route("api/[controller]")]
-public class AuthController : ControllerBase
+public class AuthController : BaseController
 {
     private readonly IAuthService _authService;
     private readonly IMediator _mediator;
@@ -32,39 +32,69 @@ public class AuthController : ControllerBase
     public async Task<ActionResult<LoginResponse>> Login([FromBody] LoginModel request)
     {
         var response = await _authService.Login(request);
-        return Ok(response);
+        return HandleResult(response);
         
     }
 
     [HttpPost]
     [Route("Register")]
-    public async Task<ActionResult<RegistrationResponse>> Register([FromBody] RegisterDto registerDto)
+    public async Task<ActionResult<Result<RegistrationResponse>>> Register([FromBody] RegisterDto registerDto)
     {
         var response = await _authService.Register(_mapper.Map<RegistrationModel>(registerDto));
         var command = new Create_UserCommand {_UserDto = _mapper.Map<Create_UserDto>(registerDto)};
-        command._UserDto.AccountId = response.UserId;
+        if (!response.Success || response.Value == null)
+            return HandleResult(response);
+
+        command._UserDto.AccountId = response.Value.UserId;
         try
         {
             var userResponse = await _mediator.Send(command);
-            return Ok(response);
+            return HandleResult(response);
         }
         catch(Exception e)
         {
             await _authService.DeleteUser(registerDto.Email);
-            response.UserId = "";
             response.Success = false;
             response.Errors.Add(e.Message);
-            return Ok(response);
+            return HandleResult(response);
         }
         
     }
 
-    // [HttpGet]
-    // [Route("Confirm")]
-    // public async Task<ActionResult<ConfirmEmailResponse>> ConfirmEmail(string token, string email)
-    // {
-    //     var response = await _authService.ConfirmEmail(token, email);
-    //     return Ok(response);
+    [HttpGet]
+    [Route("Confirm")]
+    public async Task<ActionResult<Result<string>>> ConfirmEmail(string token, string email)
+    {
+        var response = await _authService.ConfirmEmail(token, email);
+        return HandleResult(response);
         
-    // }
+    }
+
+    [HttpGet]
+    [Route("ForgotPassword")]
+    public async Task<ActionResult<Result<string>>> ForgotPassword(string email)
+    {
+        var response = await _authService.ForgotPassword(email);
+        return HandleResult(response);
+        
+    }
+
+    [HttpPost]
+    [Route("ResetPassword")]
+    public async Task<ActionResult<Result<string>>> ResetPassword([FromBody] ResetPasswordModel resetPasswordModel)
+    {
+        var response = await _authService.ResetPassword(resetPasswordModel);
+        return HandleResult(response);
+        
+    }
+
+    [HttpDelete]
+    [Route("Delete")]
+    public async Task<ActionResult<bool>> Delete(string email)
+    {
+        var response = await _authService.DeleteUser(email);
+        return Ok(response);  
+    }
+
+
 }
